[CIAS30-3702][CIAS30-3706] Pass other query params when redirecting from login page
