+++ app/containers/AnswerSessionPage/index.js
@@ -37,7 +37,6 @@ import {
 import {
   editPhoneNumberQuestionSaga,
   editUserSaga,
-  REDIRECT_QUERY_KEY,
   updateUsersTimezoneSaga,
 } from 'global/reducers/auth';
 import { resetReducer as resetAuthReducer } from 'global/reducers/auth/actions';
@@ -47,7 +46,7 @@ import {
   chatWidgetReducerKey,
   setChatEnabled,
 } from 'global/reducers/chatWidget';
-import { RoutePath } from 'global/constants';
+import { RoutePath, REDIRECT_QUERY_KEY } from 'global/constants';
 
 import { canPreview } from 'models/Status/statusPermissions';
 import { finishQuestion } from 'models/Session/QuestionTypes';
