+++ app/global/reducers/auth/sagas/logOut.js
@@ -1,11 +1,11 @@
 import { put, takeEvery, call } from 'redux-saga/effects';
 import { push } from 'connected-react-router';
 
-import { RoutePath } from 'global/constants';
+import { REDIRECT_QUERY_KEY, RoutePath } from 'global/constants';
 
 import LocalStorageService from 'utils/localStorageService';
 
-import { LOG_OUT, REDIRECT_QUERY_KEY } from '../constants';
+import { LOG_OUT } from '../constants';
 import { resetReducer } from '../actions';
 
 export function* logOut(
