+++ app/utils/router/index.ts
@@ -1,2 +1,3 @@
 export * from './matchResearchersInterventionPaths';
 export * from './parametrizeRoutePath';
+export * from './getRedirectPathFromQueryParams';
